// src/interfaces/commission.interface.ts
export interface CommissionCreate {
  productName: string | any;
  productId: string | any;
  // empName: string | any;
  // empId: string | any;
  productAmount: string | any;
  productQuantity: string | any;
  totalAmount: string | any;
  desc: string;
  type: string | any;
  commissionDate: string | any;
  commissionAmountType: string | any;
  commission: string | any;
}